+++ client/lib/ft4/crosschain/errors.ts
@@ -1,8 +1,6 @@
 export class OrchestratorError extends Error {
-  type: string;
-
-  constructor(message: string, type: string) {
+  constructor(message: string) {
     super(message);
-    this.type = type;
+    this.name = "OrchestratorError";
   }
 }