+++ client/lib/ft4/crosschain/op-functions.ts
@@ -1,8 +1,4 @@
-import {
-  SignedTransaction,
-  TransactionReceipt,
-  formatter,
-} from "postchain-client";
+import { RawGtx, TransactionReceipt, formatter } from "postchain-client";
 import { Connection } from "../types";
 import { Authenticator } from "../authentication/types";
 import {
@@ -37,7 +33,7 @@ export async function applyTransfer(
   assetId: BufferId,
   amount: Amount,
   hops: BufferId[],
-  tx: SignedTransaction,
+  tx: RawGtx,
   targetChainIndex: number,
   operationIndex: number = OP_INDEX_INIT_TRANSFER,
 ): Promise<TransactionReceipt> {