+++ client/lib/ft4/utils/entity-retriever.ts
@@ -1,6 +1,6 @@
 import { EntityRetriever, PaginatedEntity } from "./types";
 import { Connection, PagedResponse } from "../types";
-import { QueryObject, RawGtv } from "postchain-client";
+import { DictPair, QueryObject, RawGtv } from "postchain-client";
 
 export function createEntityRetriever<
   T,
@@ -8,12 +8,12 @@ export function createEntityRetriever<
   R extends PagedResponse<V> = PagedResponse<V>,
 >(
   session: Connection,
-  query: QueryObject<V>,
+  query: QueryObject<R, DictPair>,
   dataMapper: (arg: V[]) => T[],
 ): EntityRetriever<T> {
   return {
     retrieve: async (): Promise<PaginatedEntity<T>> => {
-      const res = await session.query<R>(query);
+      const res = await session.query(query);
       return {
         data: dataMapper(res.data),
         nextCursor: res.next_cursor,