+++ test/__multichain__/util/types.ts
@@ -1,6 +1,6 @@
-export interface Blockchain {
+export type Blockchain = {
   name: string;
   rid: Buffer;
   state: string;
   system: number;
-}
+};