+++ test/asset.test.ts
@@ -1,23 +1,24 @@
 import { generateAssetName, generateAssetSymbol } from "./util/util";
 import { Connection } from "../client/lib/ft4/types";
 import { createChromiaClient, getNewAsset } from "./util/blockchain-util";
-import { createConnection } from "../client/lib/ft4/ft-session";
 import { InvalidUrlError } from "../client/lib/ft4/asset/interfaces";
+import { createConnection } from "../client/lib/ft4/ft-session";
 import { Buffer } from "buffer";
 import { IClient, gtv } from "postchain-client";
 import { randomBytes } from "crypto";
 import { op } from "/ft4";
-import { adminKeyPair } from "./util/admin_user";
+import adminUser, { adminKeyPair } from "./util/admin_user";
+import { registerAsset } from "/ft4/admin/admin-op-functions";
 
 let connection: Connection;
 let client: IClient;
 
 //used only to have different issuing_brid until we have xchain
-async function registerAsset(
+async function registerAssetWithCustomBrid(
   client: IClient,
   assetName: string,
   decimals = 0,
-  blockchainRID: Buffer = randomBytes(32),
+  blockchainRid: Buffer = randomBytes(32),
 ) {
   const txn = {
     operations: [
@@ -26,7 +27,7 @@ async function registerAsset(
         assetName,
         generateAssetSymbol(),
         decimals,
-        blockchainRID,
+        blockchainRid,
         "",
       ),
     ],
@@ -58,9 +59,9 @@ describe("Asset", () => {
 
   it("can fetch paginated assets by name", async () => {
     const assetName = generateAssetName();
-    await registerAsset(client, assetName);
-    await registerAsset(client, assetName);
-    await registerAsset(client, assetName);
+    await registerAssetWithCustomBrid(client, assetName);
+    await registerAssetWithCustomBrid(client, assetName);
+    await registerAssetWithCustomBrid(client, assetName);
 
     const { data: expectedAssets, nextCursor } =
       await connection.getAssetsByName(assetName, 2);
@@ -151,11 +152,18 @@ describe("Asset", () => {
   });
 
   // Update after addding new admin functions
-  it.skip("should fail to register with invalid icon URL", async () => {
-    const invalidUrl = "not-a-valid-url";
-    await expect(
-      getNewAsset(client, "Test Asset 2", "TST2", 0, invalidUrl),
-    ).rejects.toThrow(InvalidUrlError);
+  it("should fail to register with invalid icon URL", async () => {
+    const wrapper = async () =>
+      registerAsset(
+        client,
+        adminUser().signatureProvider,
+        "Test Asset 2",
+        "TST2",
+        0,
+        "not-a-valid-url",
+      );
+
+    await expect(wrapper()).rejects.toThrow(InvalidUrlError);
   });
 
   it("should successfully register without providing icon URL", async () => {