+++ test/transaction-builder.test.ts
@@ -10,7 +10,10 @@ import {
   AuthorizationError,
   transactionBuilder,
 } from "/ft4/utils/transaction-builder";
-import { createChromiaClient } from "./util/blockchain-util";
+import {
+  anchoredHandlerCallbackParameters,
+  createChromiaClient,
+} from "./util/blockchain-util";
 import { nop } from "/ft4/utils";
 import {
   AuthDataService,
@@ -247,6 +250,7 @@ describe("Transaction Builder", () => {
     it("calls registered handler when block is anchored", async () => {
       (isBlockAnchored as jest.Mock).mockReturnValueOnce(true);
       const { authenticatorMock } = getMocks();
+      const operation = nop();
       let callback = null;
       const promise = new Promise((resolve) => {
         transactionBuilder(authenticatorMock, client)
@@ -254,14 +258,13 @@ describe("Transaction Builder", () => {
             emptyOp(),
             (callback = jest.fn().mockImplementation((op) => resolve(op))),
           )
-          .add(nop())
+          .add(operation)
           .buildAndSend();
       });
       await promise;
 
       expect(callback).toHaveBeenCalledWith(
-        emptyOp(),
-        expect.any(Buffer),
+        anchoredHandlerCallbackParameters(client, [emptyOp(), operation], 0),
         null,
       );
     });
@@ -269,6 +272,7 @@ describe("Transaction Builder", () => {
     it("calls all registered handler when block is anchored", async () => {
       (isBlockAnchored as jest.Mock).mockReturnValueOnce(true);
       const { authenticatorMock } = getMocks();
+      const operation = nop();
       let callback = null;
       let callback2 = null;
       const promise = new Promise((resolve) => {
@@ -281,19 +285,25 @@ describe("Transaction Builder", () => {
             emptyOp(),
             (callback2 = jest.fn().mockImplementation((op) => resolve(op))),
           )
-          .add(nop())
+          .add(operation)
           .buildAndSend();
       });
       await promise;
 
       expect(callback).toHaveBeenCalledWith(
-        emptyOp(),
-        expect.any(Buffer),
+        anchoredHandlerCallbackParameters(
+          client,
+          [emptyOp(), emptyOp(), operation],
+          0,
+        ),
         null,
       );
       expect(callback2).toHaveBeenCalledWith(
-        emptyOp(),
-        expect.any(Buffer),
+        anchoredHandlerCallbackParameters(
+          client,
+          [emptyOp(), emptyOp(), operation],
+          1,
+        ),
         null,
       );
     });
@@ -303,6 +313,7 @@ describe("Transaction Builder", () => {
         .mockReturnValueOnce(false)
         .mockReturnValueOnce(true);
       const { authenticatorMock } = getMocks();
+      const operation = nop();
       let callback = null;
       const promise = new Promise((resolve) => {
         transactionBuilder(authenticatorMock, client)
@@ -310,14 +321,13 @@ describe("Transaction Builder", () => {
             emptyOp(),
             (callback = jest.fn().mockImplementation((op) => resolve(op))),
           )
-          .add(nop())
+          .add(operation)
           .buildAndSend();
       });
       await promise;
 
       expect(callback).toHaveBeenCalledWith(
-        emptyOp(),
-        expect.any(Buffer),
+        anchoredHandlerCallbackParameters(client, [emptyOp(), operation], 0),
         null,
       );
     });
@@ -343,11 +353,10 @@ describe("Transaction Builder", () => {
 
       expect(callback).toHaveBeenCalledWith(
         null,
-        expect.any(Buffer),
         expect.any(AnchoringTimeoutError),
       );
     });
-    it("returns reciept without waiting for block to be anchored", async () => {
+    it("returns receipt without waiting for block to be anchored", async () => {
       const { authenticatorMock } = getMocks();
       //eslint-disable-next-line no-async-promise-executor
       const promise = new Promise(async (resolve) => {